backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
mysql-connector-python==9.7.0
python-dotenv==1.2.2